move functions to files they belong in
